refactor: move method to trait class
