+++ tests/DuskTestCase.php
@@ -2,6 +2,7 @@
 
 namespace Tests;
 
+use Exception;
 use Facebook\WebDriver\Chrome\ChromeOptions;
 use Facebook\WebDriver\Remote\DesiredCapabilities;
 use Facebook\WebDriver\Remote\RemoteWebDriver;
@@ -10,6 +11,7 @@ use Laravel\Dusk\TestCase as BaseTestCase;
 abstract class DuskTestCase extends BaseTestCase
 {
     use CreatesApplication;
+    use ApplicationUsers;
 
     /**
      * Prepare for Dusk test execution.
@@ -19,17 +21,17 @@ abstract class DuskTestCase extends BaseTestCase
      */
     public static function prepare()
     {
-      if(!getenv('CI')) {
-        static::startChromeDriver();
-      }
+        if (!getenv('CI')) {
+            static::startChromeDriver();
+        }
     }
 
-  /**
-   * Create the RemoteWebDriver instance.
-   *
-   * @return \Facebook\WebDriver\Remote\RemoteWebDriver
-   * @throws \Exception
-   */
+    /**
+     * Create the RemoteWebDriver instance.
+     *
+     * @return RemoteWebDriver
+     * @throws Exception
+     */
     protected function driver()
     {
         $options = (new ChromeOptions)->addArguments([
@@ -40,11 +42,11 @@ abstract class DuskTestCase extends BaseTestCase
         ]);
 
         $capabilities = DesiredCapabilities::chrome()->setCapability(
-          ChromeOptions::CAPABILITY, $options
+            ChromeOptions::CAPABILITY, $options
         );
 
         return retry(5, function () use ($capabilities) {
-          return RemoteWebDriver::create('http://localhost:9515', $capabilities, 60000, 60000);
+            return RemoteWebDriver::create('http://localhost:9515', $capabilities, 60000, 60000);
         }, 50);
     }
 }
