+++ tests/TestCase.php
@@ -10,16 +10,5 @@ use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 abstract class TestCase extends BaseTestCase
 {
     use CreatesApplication;
-
-    protected function createUserWithPermissions(array $permissions)
-    {
-        (new RolesAndPermissionsSeeder())->run();
-        Role::create(['name' => 'test-role'])
-            ->givePermissionTo($permissions);
-
-        $user = User::factory()->make();
-        $user->assignRole(['test-role']);
-        $user->save();
-        return $user;
-    }
+    use ApplicationUsers;
 }
